requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.40.0
peft>=0.10.0
accelerate>=0.28.0
datasets>=2.19.0
huggingface_hub>=0.22.0

# Medical imaging
monai>=1.3.0
nibabel>=5.2.0
pydicom>=2.4.0
pylidc>=0.2.3
SimpleITK>=2.3.0
scikit-image>=0.22.0

# Training utilities
einops>=0.7.0
sentencepiece>=0.2.0
bitsandbytes>=0.43.0
deepspeed>=0.14.0

# Evaluation
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0
scikit-learn>=1.4.0

# Retrieval
faiss-cpu>=1.8.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0.1
Pillow>=10.3.0
matplotlib>=3.8.0
